Add identity hash vector
This is an old matcher from the original rematch project.
This aims to generate a unique hash for a function based on its exact binary
instructions, only ignoring reference offsets (as those might change based on
the function's location in the file even for exact copies).

Reference offsets are ignored by only taking the first instruction byte, by
operating under the assumption the first byte holds the type of reference
(call, jump, conditional jump) and subsequent bytes the actual offsets.

Signed-off-by: Nir Izraeli <nirizr@gmail.com>